added router, structural changes
