+++ src/components/NavBar.css
@@ -1,0 +1,40 @@
+/* Navbar Content */
+
+.navBar {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 3rem;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    padding: 0 2rem;
+    padding-bottom: 1rem;
+    border-bottom: 1px solid white;
+    z-index: 100;
+  }
+
+  .navBar h2 a {
+    text-decoration: none;
+    color: white;
+  }
+  
+  .nav {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+    margin-right: 3rem;
+    padding-top: 0.5rem;
+    list-style: none;
+  }
+  
+  .nav li {
+    margin: 0 1rem;
+  }
+  
+  .nav li a {
+    text-decoration: none;
+    color: white;
+  }
+
